requirements.txt
django==3.1.6
django-filter==2.4.0  
djangorestframework==3.12.2 
Markdown==3.3.3 